+++ app/Main.hs
@@ -1,3 +1,12 @@
+{-|
+Module      : Main
+Description : The main module for the wai-devel executable.
+Copyright   : (c) 2015 Njagi Mwaniki
+License     : MIT
+Maintainer  : njagi@urbanslug.com
+Stability   : experimental
+Portability : POSIX
+-}
 module Main where
 
 import DevelMain